ml-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19
tensorflow-cpu==2.18.0
numpy==1.26.4
Pillow==10.4.0
pydantic==2.10.3
scikit-learn==1.5.2
